+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include README.md
+include tech_sage/adress_book_1.pkl
