+++ tech_sage.egg-info/entry_points.txt
@@ -1,0 +1,2 @@
+[console_scripts]
+tech_sage = tech_sage.main:main
