+++ tech_sage.egg-info/requires.txt
@@ -1,0 +1,2 @@
+prompt_toolkit
+rich
